requirements.txt
reportlab>=4.0
matplotlib>=3.7
numpy>=1.24
apscheduler>=3.10
python-dotenv>=1.0
flask>=3.0
cryptography>=42.0
